+++ hutool-db/src/main/java/cn/hutool/db/Db.java
@@ -1,5 +1,0 @@
-package cn.hutool.db;
-
-public class Db {
-
-}